Fake `mktemp` when calling Simon's build script
Simon's `build.sh` uses `ccache` in an attempt to speed
up builds. However, it does not really work, because the
build path is different each time due to use of `mktemp`
and `ccache` uses file path when computing hash.

This commit provides fake implementation of `mktemp` that
actually does not create unique and random temp directory,
assuming:

 1) build uses custom temp directory, not global `/tmp`
 2) mktemp is used only once for create build directory
    for GDB.

As of now, both assumptions hold. Hope it will stay like
this.