Update `test.sh`
Up to now test script used shell script from lttng-ci [1] to build, run
and process tests. Recently the script has been removed in favour of
a groovy script (see commit 661534f [2]).

To keep things going, this commit updates `test.sh` to:

 * use forked lttng-ci scripts
 * checkout last version that contained freestyle bash script
   to build and test GDB.

In the long run it'd be better to migrate to the groovy script but
this is considerable work so for now, I decided to take the easier
way.

[1]: https://github.com/lttng/lttng-ci
[2]: https://github.com/lttng/lttng-ci/commit/661534f5a85cb88fcaba92b85149ce8c90f30e34