Move required scripts from lttng-ci repo here
Previous commit switched to use fixed version of LTTng CI scripts
because upstream switched from using shell script to groovy script.

This commit goes further and adds copy of these scripts to this repo.

The original idea of using scripts from that repo was to keep in sync
with what other GDB developers use but after the switch to groovy this
is passe.

Moreover we need to tweak the scripts a bit to add more know failures
and/or tweak C/C++ compiler flags, mainly because my build hosts differ
from those at EfficiOS. This was done in really hacky way by maintaing
a "patch" to upstream scripts.

Because of all this, I recon it is just easier to maintain my own
copy of these script.