Allow reading if known failures from externally provided file
This commit changes build.sh to *also* add contents of
`${WORKSPACE}/known-failures-*` to list of known failures.

This allows CI jobs to download and drop in failures of some
reference build ("master" build) so the `build.sh` would only
report failures that happened in this version but not in the
reference version.

Useful for "try" builds of development branches to see if they
break anything that is not broken in "master" (sure, ideally
"master" should be clean but it is not at the moment).